+++ ChangeLog.md
@@ -103,6 +103,8 @@ These artifacts include extensions for the types available in the latter JDKs, s
 
 #### Intention actions, inspections and quickfixes
 
+- [`KT-9490`](https://youtrack.jetbrains.com/issue/KT-9490) Convert receiver to parameter: use template instead of the dialog
+
 ##### New features
 
 - [`KT-11525`](https://youtrack.jetbrains.com/issue/KT-11525) Implement "Create type parameter" quickfix